Run integration tests in Docker container.
This builds and run a Docker container when doing integration tests
(uploading and building extensions). By using a Docker contianer instead of the
running host makes the test more production like.

Also, this updates how metrics are reported. It now sends the right environment
parameter so that we can distinguish between local and production servers.
