+++ server/scripts/stop-test-server.sh
@@ -1,0 +1,4 @@
+#!/usr/bin/env bash
+
+docker logs extender
+docker stop extender
